+++ Rerecipe/src/de/rerecipe/model/Search.java
@@ -6,9 +6,9 @@ public class Search {
 
 	public static class EnteredIngredient {
 		private String name;
-		private int amount;
+		private double amount;
 
-		public EnteredIngredient(String name, int amount) {
+		public EnteredIngredient(String name, double amount) {
 			this.name = name;
 			this.amount = amount;
 		}
@@ -17,7 +17,7 @@ public class Search {
 			return name;
 		}
 
-		public int getAmount() {
+		public double getAmount() {
 			return amount;
 		}
 
@@ -27,10 +27,10 @@ public class Search {
 	private List<String> filterOptions;
 	private String orderBy;
 	private int start;
-	private int amount;
+	private double amount;
 
 	public Search(List<EnteredIngredient> ingredients, List<String> filterOptions,
-			String orderBy, int start, int amount) {
+			String orderBy, int start, double amount) {
 		this.ingredients = ingredients;
 		this.filterOptions = filterOptions;
 		this.orderBy = orderBy;
@@ -54,7 +54,7 @@ public class Search {
 		return start;
 	}
 
-	public int getAmount() {
+	public double getAmount() {
 		return amount;
 	}
